requirements.txt
streamlit
diffusers
transformers
Pillow